requirements.txt
email_validator==2.2.0
Flask~=3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy
Flask-WTF==1.2.2
Jinja2
psycopg2-binary==2.9.10
pytest
python-dateutil
SQLAlchemy
Werkzeug
WTForms
greenlet
idna
iniconfig
itsdangerous~=2.2.0
Jinja2~=3.1.4
Mako==1.3.8
MarkupSafe
packaging
pluggy
psycopg2-binary==2.9.10
pycparser
pytest~=8.3.4
python-dateutil
setuptools==75.1.0
six
sql==2022.4.0
SQLAlchemy~=2.0.38
typing_extensions
Werkzeug~=3.1.3
wheel==0.44.0
WTForms~=3.2.1
google-auth-oauthlib~=1.2.2
google-api-python-client~=2.169.0
protobuf~=6.30.2
python-dotenv~=0.21.0
colorlog>=5.0.0
paho-mqtt
psycopg2
